semester table refactor - skeleton
