+++ packages/sogrim-app/package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
     "@mui/icons-material": "^5.0.4",
     "@mui/material": "^5.0.3",
     "@testing-library/jest-dom": "^5.11.4",
@@ -27,6 +29,7 @@
     "react-query": "^3.31.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
+    "react-table": "^7.7.0",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
@@ -56,6 +59,7 @@
   },
   "devDependencies": {
     "@types/js-cookie": "^3.0.0",
-    "@types/react-router-dom": "^5.3.2"
+    "@types/react-router-dom": "^5.3.2",
+    "@types/react-table": "^7.7.9"
   }
 }
