+++ packages/sogrim-app/src/components/Pages/SemestersPage/SemesterTable/SemesterTableBody.tsx
@@ -1,6 +1,7 @@
 import { TableBody } from "@mui/material";
-import { SemesterTableRow } from "./SemesterTableRow";
 import { RowData } from "./SemesterTabsConsts";
+import { TableRow } from "@mui/material";
+import { ReadOnlyRow } from "./SemesterTableRow/ReadOnlyRow";
 
 interface SemesterTableBodyProps {
     tableRows: RowData[];
@@ -12,16 +13,33 @@ export const SemesterTableBody: React.FC<SemesterTableBodyProps> = ({
     tableRows,
     handleSave,
     semester,
-}) => {
+}) => {    
+
+
+    const handleEdit = () => {
+    }; 
+
+    const handleDelete = () => {
+    };
+
+    const handleCancel = () => {
+    };
 
     return (
         <TableBody>         
               {(tableRows)                
                 .map((row, index) => {                  
                   const labelId = `enhanced-table-checkbox-${index}`;
-                  console.log('row!!!', row);
                   return (
-                      <SemesterTableRow row={row} labelId={labelId} handleSave={handleSave} semester={semester} key={index} />                   
+                    <TableRow
+                    hover            
+                    role="checkbox"            
+                    tabIndex={-1}
+                    key={row.name}            
+                    sx={{ width: '1200px'}}
+                    >                      
+                        <ReadOnlyRow row={row} labelId={labelId} handleEdit={handleEdit} handleDelete={handleDelete}/>                       
+                    </TableRow>             
                   )
                 })}              
         </TableBody>
