+++ packages/sogrim-app/src/components/Pages/SemestersPage/SemestersPage.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { observer } from 'mobx-react-lite';
 import { SemesterTabs } from './SemesterTabs';
+
 interface SemestersTabProps {
 }
 
 const SemestersPageComp: React.FC<SemestersTabProps> = () => {
     return (        
-        <SemesterTabs />        
+       <SemesterTabs />
     );
 };
